chore: rename packages
