+++ .eslintrc.js
@@ -1,5 +1,5 @@
 /** @type {import('eslint').Linter.Config} */
 module.exports = {
   root: true,
-  ...require('./packages/eslint'),
+  ...require('./packages/eslint-config'),
 };
