+++ packages/prettier-config/index.js
