+++ tsconfig.json
@@ -1,6 +1,6 @@
 {
   "$schema": "https://json.schemastore.org/tsconfig",
-  "extends": "./packages/tsconfig/tsconfig.json",
+  "extends": "./packages/typescript-config/tsconfig.json",
   "compilerOptions": {
     "noEmit": true
   }
